Added schedule backgroud, moved ScheduleTimer to its own directory
