+++ apps/site/src/app/schedule/Header/Header.module.scss
@@ -1,8 +1,8 @@
 .header {
-	margin-top: 200px;
 	height: auto;
 	display: flex;
 	flex-direction: column;
-	gap: 50px;
 	align-items: center;
+
+	gap: 50px;
 }
