+++ apps/site/src/app/schedule/Schedule.tsx
@@ -1,9 +1,22 @@
+import Image from "next/image";
+
 import { getSchedule } from "./getSchedule";
+import ScheduleTimer from "./ScheduleTimer/ScheduleTimer";
+
+import background from "@/assets/images/resources-background.svg";
 
 import styles from "./Schedule.module.scss";
-import ScheduleTimer from "./ScheduleTimer";
 
 export default async function Schedule() {
 	const schedule = await getSchedule();
-	return <ScheduleTimer schedule={schedule}></ScheduleTimer>;
+	return (
+		<>
+			<Image
+				src={background}
+				alt="cloud background"
+				className={styles.background}
+			/>
+			<ScheduleTimer schedule={schedule} />
+		</>
+	);
 }
