+++ apps/site/src/app/schedule/ScheduleTimer/ScheduleTimer.module.scss
@@ -1,0 +1,3 @@
+.timer {
+	padding-top: 200px;
+}
